fix(nextjs): copy helpers.js in prebuild + retarget moved imports
CI test-nextjs-f7 build was failing:
  Attempted import error: 'expandTemplatesSync' is not exported from
  '#utils/hydra' (imported as 'expandTemplatesSync').

hydra-nextjs vendors a copy of hydra.js into src/utils/ via the
copy:hydra prebuild script (Next.js can't resolve workspace packages
when running with --ignore-workspace as CI does). After 95d5652 +
03c6bad moved 21+ Plone helpers from hydra-js to @volto-hydra/helpers,
the vendored hydra.js no longer exports them.

Fix:
  - copy:hydra script now copies BOTH packages/hydra-js/hydra.js and
    packages/helpers/index.js into src/utils/.
  - Added "#utils/helpers" import alias to package.json next to
    "#utils/hydra".
  - .gitignore covers src/utils/helpers.js alongside hydra.js.
  - src/components/BlocksList/BlocksList.js: all 5 names in its
    import line are moved helpers, swap the whole import from
    "#utils/hydra" to "#utils/helpers".
  - src/utils/api.js: loadTemplates moved, swap import.
  - PageClient.js (initBridge), Menu.js / RecursiveMenuItem.js
    (getTokenFromCookie) keep importing from "#utils/hydra" -
    those names stay in @volto-hydra/hydra-js.

Verified locally:
  cd examples/hydra-nextjs
  pnpm install --ignore-workspace
  pnpm run build:test     -> succeeds, route compiled
