+++ examples/hydra-nextjs/.gitignore
@@ -131,5 +131,6 @@ dist
 
 certificates
 
-# Copied from packages/hydra-js at build time
+# Copied from packages/{hydra-js,helpers} at build time (see copy:hydra script)
 src/utils/hydra.js
+src/utils/helpers.js
