+++ examples/hydra-nextjs/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "copy:hydra": "pnpm -F @volto-hydra/hydra-js build && cp ../../packages/hydra-js/hydra.js ./src/utils/hydra.js",
+    "copy:hydra": "pnpm -F @volto-hydra/hydra-js build && cp ../../packages/hydra-js/hydra.js ./src/utils/hydra.js && cp ../../packages/helpers/index.js ./src/utils/helpers.js",
     "predev": "pnpm run copy:hydra",
     "dev": "next dev -p 3002",
     "dev:test": "pnpm run copy:hydra && next dev -p 3007",
@@ -18,6 +18,7 @@
   },
   "imports": {
     "#utils/hydra": "./src/utils/hydra.js",
+    "#utils/helpers": "./src/utils/helpers.js",
     "#utils/*": "./src/utils/*.js"
   },
   "dependencies": {
